% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPvals.R
\name{compPvals}
\alias{compPvals}
\title{compute p-values}
\usage{
compPvals(MD, dist.method = "MMD", diff.method = "MMD.locfit")
}
\arguments{
\item{MD}{DBAmmd Object. This Object can be created using \code{DBAmmd()}.}

\item{dist.method}{which method should be used to determine distances
between samples. Currently only Maximum Mean Discrepancy (MMD)
and Kolmogorov-Smirnov (KS) implemented.
(DEFAULT: 'MMD')}

\item{diff.method}{(DEFAULT: 'MMD.locfit')}
}
\value{
DBAmmd object with updated Contrast Dists
}
\description{
This function determines peak-specific p-values based on distances between
sample histograms.
}
\examples{

## Example using a small data set provided with this package:

data("MMD")
MMD.1 <- compPvals(MMD)

}
\seealso{
\code{\link{DBAmmd}}, \code{\link{plotDists}},\code{\link{compDists}}
}

