% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBAmmd-Accessors.R
\docType{methods}
\name{DBAmmd-Accessors}
\alias{Contrast,DBAmmd-method}
\alias{Counts,DBAmmd-method}
\alias{DBAmmd-Accessors}
\alias{Dists,DBAmmd-method}
\alias{Genome,DBAmmd-method}
\alias{Hists,DBAmmd-method}
\alias{Reads,DBAmmd-method}
\alias{Regions,DBAmmd-method}
\alias{Samples,DBAmmd-method}
\alias{metaData,DBAmmd-method}
\alias{numPeaks,DBAmmd-method}
\alias{numSamples,DBAmmd-method}
\alias{setContrast,DBAmmd-method}
\alias{setRegions,DBAmmd-method}
\title{Extract data from DBAmmd objects}
\usage{
\S4method{Genome}{DBAmmd}(x)

\S4method{Samples}{DBAmmd}(x)

\S4method{numPeaks}{DBAmmd}(x)

\S4method{numSamples}{DBAmmd}(x)

\S4method{metaData}{DBAmmd}(x)

\S4method{Regions}{DBAmmd}(x)

\S4method{Reads}{DBAmmd}(x, whichPos = NULL)

\S4method{Counts}{DBAmmd}(x, whichCounts = "T")

\S4method{Hists}{DBAmmd}(x, whichPos)

\S4method{Dists}{DBAmmd}(x, dist.method = NULL)

\S4method{Contrast}{DBAmmd}(x, whichContrast = 1)

\S4method{setRegions}{DBAmmd}(x, Regions)

\S4method{setContrast}{DBAmmd}(x, contrast)
}
\arguments{
\item{x}{a DBAmmd Object. This Object can be created using \code{DBAmmd()}.
As MetaData it needs to contain the path to the data directory and the
name of a sampleSheet csv file. Additionally, the Regions of interests must
be provided as GRanges Object.}

\item{whichPos}{specifies which relative positions of mapped fragments
should to be considered.
Can be one of: 'Left.p', 'Right.p': Start and end positions of fragments
mapping to positive strand, respectively.
('Right.p' is not available for single-end reads)
'Left.n','Right.n': Start and end positions of fragments
mapping to negative strand, respectively.
('Left.n' is not available for single-end reads)
Additionally inferred positions: 'Center.n','Center.p','Center','Left','Right'.}

\item{whichCounts}{can be 'T': total counts, or
'p','n': counts of reads mapping to positive, negative strand, respectively}

\item{dist.method}{which method should be used to determine distances
between samples. Currently only Maximum Mean Discrepancy (MMD)
and Kolmogorov-Smirnov (KS) implemented.
(DEFAULT: 'MMD')}

\item{whichContrast}{(DEFAULT: 1)}

\item{Regions}{GRanges Object}

\item{contrast}{how to set contrast, either 'byCondition', or 'byTissue'.
contrast can also be manually set (see vignette for details.)}
}
\value{
\code{Genome(x)} returns

\code{Samples(x)} returns

\code{numPeaks(x)} returns

\code{numSamples(x)} returns

\code{metaData(x)} returns

\code{Regions(x)} returns

\code{Reads(x,whichPos)} returns

\code{Counts(x,whichCounts)} returns

\code{Hists(x,whichPos)} returns

\code{Dists(x,dist.method)} returns

\code{Contrast(x,whichContrast)} returns

\code{setRegions(x,Regions)} returns a DBAmmd Object with Regions set

\code{setContrast(x,contrast)} returns a DBAmmd Object
with a contrast set
}
\description{
This help file describes different ways to access the slots and values
contained in \code{\link{DBAmmd}} objects.
}
\examples{
data("MMD")

Samples(MMD)
Genome(MMD)
numPeaks(MMD)
numSamples(MMD)
metaData(MMD)
R <- Regions(MMD)
Pos <- Reads(MMD, whichPos='Center')
C <- Counts(MMD)
H <- Hists(MMD, whichPos='Center')
D <- Dists(MMD)
C1 <- Contrast(MMD)

}

