% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBAmmd-Class.R
\docType{class}
\name{DBAmmd-class}
\alias{DBAmmd}
\alias{DBAmmd-class}
\title{Class DBAmmd}
\value{
DBAmmd Object
}
\description{
Class \code{DBAmmd} defines a container for differential binding analysis
using MMDiff2. For this class a number of methods is foreseen, among which
accessors for every slot.
As MetaData it needs to contain the path to the data directory and the
name of a sampleSheet csv file. Additionally, the Regions of interests must
be provided as GRanges Object.
}
\note{
For this class,
}
\section{Slots}{

\describe{
\item{\code{MetaData}:}{MetaData}
\item{\code{rowRanges}:}{GRanges list object containing Regions of Interests
(Peaks)}
\item{\code{Reads}:}{List containing positions of mapped reads, i.e. exact
start and end positions of mapped fragments, not coverage. In the case of
single-end reads, the left most postions of fragments mapping to the positive
strands and the right most positions of fragments
mapping to the negative strands are stored in "Left.p" and "Right.n".
Use \code{getPeakReads} to fill this slot and \code{estimateFragmentCenters}
to add the (estimated) positions of fragment centers.}
\item{\code{RawTotalCounts}:}{ m x n matrix containing total counts of reads
mapping
to m peaks and in n samples (including input samples)}
\item{\code{RawCounts.p}:}{ m x n matrix containing counts of reads mapping
to positive (forward) strand}
\item{\code{RawCounts.n}:}{ m x n matrix containing counts of reads mapping
to negative (forward) strand}
\item{\code{Hists}:}{ For each Peak matrix}
\item{\code{DISTs}:}{ for each Peak computed distances between pairs of
samples}
\item{\code{mCounts}:}{ (for internal use only, counts averaged over pairs of
samples)}
\item{\code{Contrasts}:}{ 'list'}
}
}
\author{
Gabriele Schweikert
}
\seealso{
\code{\link{DBAmmd-Accessors}}
}

