% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeak.R
\name{plotPeak}
\alias{plotPeak}
\title{plot Peak}
\usage{
plotPeak(MD, Peak.id, Sample.ids = NULL, NormMethod = NULL,
  plot.input = FALSE, whichPos = "Center", which.contrast = NULL,
  Motifs = NULL, Motifcutoff = "80\%", anno = NULL, xaxt = NULL,
  xlim = NULL, ylim = NULL)
}
\arguments{
\item{MD}{DBAmmd Object. This Object can be created using \code{DBAmmd()}.}

\item{Peak.id}{Peak id (coresponding to names of Regions(MD))}

\item{Sample.ids}{which samples to draw. IF NULL all samples are drawn.
(DEFAULT: NULL)}

\item{NormMethod}{wether to apply normailzation factors.
currently no NormMetimplemented (DEFAULT: None)}

\item{plot.input}{wether to plot input controls (DEFAULT: TRUE)}

\item{whichPos}{specifies which relative positions of mapped fragments
should to be considered.
Can be one of: 'Left.p', 'Right.p': Start and end positions of fragments
mapping to positive strand, respectively.
('Right.p' is not available for single-end reads)
'Left.n','Right.n': Start and end positions of fragments
mapping to negative strand, respectively.
('Left.n' is not available for single-end reads)
Additionally inferred positions: 'Center.n','Center.p','Center','Left','Right'.}

\item{which.contrast}{(DEFAULT: 1)}

\item{Motifs}{TF binding sites (DEFAULT: NULL)}

\item{Motifcutoff}{(Default: "80\%")}

\item{anno}{(Default: NULL)}

\item{xaxt}{(Default: NULL)}

\item{xlim}{(Default: NULL)}

\item{ylim}{(Default: NULL)}
}
\description{
This function plots a peak
}
\examples{
dev.off()
data("MMD")
plotPeak(MMD,Peak.id='6',plot.input=FALSE)

# add annotation (Overlapping genes)
load(system.file("data/mm9-Genes.RData", package="MMDiff2"))
GR <- list(UCSCKnownGenes = GR)
plotPeak(MMD, Peak.id='6', plot.input = FALSE, anno=GR)

# add TF binding sites
library('MotifDb')
plot_motifs <- query(query(MotifDb, 'Mmusculus'), 'E2F')
plotPeak(MMD, Peak.id='6', plot.input = FALSE,
       Motifs=plot_motifs,Motifcutoff="80\%")

# split peaks by contrast
plotPeak(MMD, Peak.id='6', plot.input = FALSE, which.contrast=1,
       Motifs=plot_motifs,Motifcutoff="80\%",anno=GR)


}

