% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportResults.R
\name{reportResults}
\alias{reportResults}
\title{reportResults}
\usage{
reportResults(MD, diff.method = "MMD.locfit", th = 0.1,
  which.contrast = 1, rm.oulier = TRUE, bUsePval = FALSE)
}
\arguments{
\item{MD}{DBAmmd Object. This Object can be created using \code{DBAmmd()}.}

\item{diff.method}{(DEFAULT: 'MMD.locfit')}

\item{th}{significance threshold to highlight peaks (DEFAULT: 0.1)}

\item{which.contrast}{(DEFAULT: 1)}

\item{rm.oulier}{(DEFAULT: TRUE)}

\item{bUsePval}{if TRUE highlight significant peaks according to their
p-values, else use FDR (DEFAULT: FALSE)}
}
\description{
retrieve Results of differential binding analysis
}
\examples{

data("MMD")
res <- reportResults(MMD)

}

