% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyMMDiff2.R
\name{runShinyMMDiff2}
\alias{runShinyMMDiff2}
\title{Shiny Application for interactive visualization of MMD,GMD and
Pearson Difference as well as plotting peaks}
\usage{
runShinyMMDiff2(MD, con = NULL)
}
\arguments{
\item{MD}{DBAmmd object}

\item{con}{(DEFAULT: NULL)}
}
\description{
Shiny Application for interactive visualization of MMD,GMD and
Pearson Difference as well as plotting peaks
}
\examples{
 if(interactive()){
load(system.file("data/MMD.RData", package="MMDiff2"))
runShinyMMDiff2(MMD)
}

}

