% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{metaData}
\alias{Contrast}
\alias{Counts}
\alias{Dists}
\alias{Genome}
\alias{Hists}
\alias{Reads}
\alias{Regions}
\alias{Samples}
\alias{metaData}
\alias{numPeaks}
\alias{numSamples}
\alias{setContrast}
\alias{setRegions}
\title{Generics for DBAmmd-Class}
\usage{
metaData(x, ...)

Regions(x, ...)

Reads(x, ...)

Counts(x, ...)

Hists(x, ...)

Dists(x, ...)

Contrast(x, ...)

numPeaks(x, ...)

numSamples(x, ...)

Samples(x, ...)

Genome(x, ...)

setRegions(x, ...)

setContrast(x, ...)
}
\arguments{
\item{x}{a DBAmmd Object. An empty instance can be created using \code{DBAmmd()}.
(See \code{\link{DBAmmd-class}} for more details.)}

\item{...}{additional parameters}
}
\description{
Generics for DBAmmd-Class
}

